TareaDosEntregaCuatroVUno

+++ src/main/java/com/github/ignacioalbornoz/finalreality/model/character/IEnemy.java
@@ -13,5 +13,5 @@ public interface IEnemy extends ICharacter{
      */
     int getWeight();
 
-    float getDamage();
+    int getDamage();
 }
